Keep the error modal for what the visitor can act on
`suppressErrorModal` was passed at the top level of the request, where the
RequestManager does not allow keys of our own: the request was turned down
before it was sent. Only debug builds check, so the one caller -- filling a
new application in from the last one -- worked in production and quietly did
nothing everywhere else, with no test to say so.

Custom data belongs in `options`, which the RequestManager does allow.

Nobody being signed in joins it: it is not something to put in front of a
visitor, who is shown the way in instead.
