+++ web/app/components/submission-form/files.gts
@@ -45,7 +45,7 @@ export default class SubmissionFormFilesComponent extends Component<Signature> {
     try {
       const { content } = await this.requestManager.request<LastSubmitted>({
         url: '/submissions/last_submitted',
-        suppressErrorModal: true,
+        options: { suppressErrorModal: true },
       });
 
       last = content.submission;
