+++ web/app/request-handlers/types.d.ts
@@ -3,6 +3,5 @@ import '@warp-drive/core/types/request';
 declare module '@warp-drive/core/types/request' {
   interface RequestInfo {
     data?: Record<string, unknown>;
-    suppressErrorModal?: boolean;
   }
 }
